refactor: extract resolver types to types.ts (#25)
