+++ src/internal/resolver/types.ts
@@ -1,0 +1,56 @@
+import type { HoconValue } from '../../value.js'
+
+// ---- Internal placeholder types ----
+export type SubstPlaceholder = {
+  _kind: 'subst-placeholder'
+  path: string
+  optional: boolean
+  line: number
+  col: number
+  prefixLen: number  // 0 for normal, >0 for relativized (number of prefix segments)
+}
+export type ConcatPlaceholder = {
+  _kind: 'concat-placeholder'
+  nodes: ResolverValue[]
+}
+export type AppendPlaceholder = {
+  _kind: 'append-placeholder'
+  existing: ResolverValue
+  elem: ResolverValue
+}
+export type ResObj = {
+  _kind: 'res-obj'
+  fields: Map<string, ResolverValue>
+  priorValues: Map<string, ResolverValue>
+}
+
+export type ResolverValue = HoconValue | SubstPlaceholder | ConcatPlaceholder | AppendPlaceholder | ResObj
+
+export type ResolveOptions = {
+  env: Record<string, string>
+  baseDir: string | undefined
+  readFileSync: (filePath: string) => string
+  readFile?: (filePath: string) => Promise<string>
+  includeStack?: string[]
+}
+
+// Track parser-inserted separator whitespace values without leaking _separator
+// into the public HoconValue type. Uses WeakSet so values can be GC'd normally.
+export const separatorValues = new WeakSet<HoconValue>()
+
+export function isSubst(v: ResolverValue): v is SubstPlaceholder {
+  return (v as SubstPlaceholder)._kind === 'subst-placeholder'
+}
+export function isConcat(v: ResolverValue): v is ConcatPlaceholder {
+  return (v as ConcatPlaceholder)._kind === 'concat-placeholder'
+}
+export function isAppend(v: ResolverValue): v is AppendPlaceholder {
+  return (v as AppendPlaceholder)._kind === 'append-placeholder'
+}
+export function isResObj(v: ResolverValue): v is ResObj {
+  return (v as ResObj)._kind === 'res-obj'
+}
+
+export function makeResObj(): ResObj {
+  return { _kind: 'res-obj', fields: new Map(), priorValues: new Map() }
+}
